ai-engine/nlp-processor/requirements.txt
# NLP Processor Dependencies
# Python 3.10+

# LangChain
langchain>=0.1.0
langchain-anthropic>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.1.0

# Anthropic Claude
anthropic>=0.18.0

# FastAPI
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.0.0

# HTTP client
httpx>=0.25.0,<0.27.0
aiohttp>=3.9.0

# JSON processing
orjson>=3.9.0

# Utilities
python-dotenv>=1.0.0
tenacity>=8.2.0

# Redis (async)
redis>=5.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Type checking
mypy>=1.7.0
